refactor: déplacer StepsSection de src/components vers src/app/(components)/steps
- Déplacer tous les composants StepsSection vers le dossier (components)
- Mettre à jour les imports pour utiliser les chemins relatifs
- Intégrer le composant de base directement dans StepsSection.tsx
- Supprimer l'ancien dossier src/components/StepsSection
- Maintenir la compatibilité avec l'internationalisation
