+++ src/app/(components)/steps/ImageWithRoundedCorner.tsx
