+++ src/app/(components)/steps/SectionHeader.tsx
